--- src/hwenc/README.md
@@ -1,3 +1,0 @@
-# Sora C++ SDK から持ってきたハードウェアエンコーダ/デコーダを配置するディレクトリ
-
-sora-cpp-sdk からコピペしやすいように、ディレクトリを独立させました。
