+++ src/sora-cpp-sdk/README.md
@@ -1,0 +1,1 @@
+# Sora C++ SDK から持ってきたソースコードを配置するディレクトリ
