+++ src/sora-cpp-sdk/include/sora/cuda_context.h
@@ -3,9 +3,13 @@
 
 #include <memory>
 
+namespace sora {
+
 // <cuda.h> に依存せずに CUDA のコンテキストを表す
 class CudaContext {
  public:
+  // CUDA コンテキスト生成する。
+  // CUDA に対応していないプラットフォームでは nullptr を返す。
   static std::shared_ptr<CudaContext> Create();
   void* Context() const;
 
@@ -19,5 +23,7 @@ enum class CudaVideoCodec {
   VP9,
   JPEG,
 };
+
+}  // namespace sora
 
 #endif
