+++ src/sora-cpp-sdk/include/sora/dyn/dyn.h
@@ -8,7 +8,7 @@
 
 #if defined(_WIN32)
 #include <windows.h>
-#elif defined(__linux__)
+#else
 // Linux
 #include <dlfcn.h>
 #endif
@@ -28,7 +28,7 @@ class DynModule {
   typedef void* module_ptr_t;
 #endif
 
-  bool IsLoadable(const char* name) {
+  static bool IsLoadable(const char* name) {
 #if defined(_WIN32)
     module_ptr_t module = LoadLibraryA(name);
     if (module == nullptr) {
@@ -36,15 +36,13 @@ class DynModule {
     }
     FreeLibrary(module);
     return true;
-#elif defined(__linux__)
+#else
     module_ptr_t module = dlopen(name, RTLD_LAZY);
     if (module == nullptr) {
       return false;
     }
     dlclose(module);
     return true;
-#else
-    return false;
 #endif
   }
 
@@ -55,10 +53,8 @@ class DynModule {
     }
 #if defined(_WIN32)
     module_ptr_t module = LoadLibraryA(name);
-#elif defined(__linux__)
-    module_ptr_t module = dlopen(name, RTLD_LAZY);
 #else
-    module_ptr_t module = nullptr;
+    module_ptr_t module = dlopen(name, RTLD_LAZY);
 #endif
     if (module == nullptr) {
       return nullptr;
@@ -75,10 +71,8 @@ class DynModule {
     }
 #if defined(_WIN32)
     return ::GetProcAddress(module, name);
-#elif defined(__linux__)
-    return dlsym(module, name);
 #else
-    return nullptr;
+    return dlsym(module, name);
 #endif
   }
 
@@ -88,9 +82,8 @@ class DynModule {
       if (p != nullptr) {
 #if defined(_WIN32)
         FreeLibrary(p);
-#elif defined(__linux__)
-        ::dlclose(p);
 #else
+        ::dlclose(p);
 #endif
       }
     }
@@ -118,8 +111,8 @@ class DynModule {
     auto f =                                                                   \
         (func_type)DynModule::Instance().GetFunc(soname, DYN_STRINGIZE(func)); \
     if (f == nullptr) {                                                        \
-      std::cerr << "Failed to load function " DYN_STRINGIZE(func) " in "       \
-                << soname << std::endl;                                        \
+      std::cerr << "Failed to GetFunc: " << DYN_STRINGIZE(func)                \
+                << " soname=" << soname << std::endl;                          \
       exit(1);                                                                 \
     }                                                                          \
     return f(args...);                                                         \
