+++ src/sora-cpp-sdk/include/sora/hwenc_nvcodec/nvcodec_v4l2_capturer.h
@@ -1,9 +1,10 @@
-#ifndef NVCODEC_V4L2_CAPTURER_H_
-#define NVCODEC_V4L2_CAPTURER_H_
+#ifndef SORA_HWENC_NVCODEC_NVCODEC_V4L2_CAPTURER_H_
+#define SORA_HWENC_NVCODEC_NVCODEC_V4L2_CAPTURER_H_
 
-#include "v4l2_video_capturer/v4l2_video_capturer.h"
+#include "nvcodec_decoder_cuda.h"
+#include "sora/v4l2/v4l2_video_capturer.h"
 
-#include "./nvcodec_decoder_cuda.h"
+namespace sora {
 
 struct NvCodecV4L2CapturerConfig : V4L2VideoCapturerConfig {
   NvCodecV4L2CapturerConfig(const V4L2VideoCapturerConfig& config) {
@@ -15,15 +16,13 @@ struct NvCodecV4L2CapturerConfig : V4L2VideoCapturerConfig {
 class NvCodecV4L2Capturer : public V4L2VideoCapturer {
  public:
   static rtc::scoped_refptr<V4L2VideoCapturer> Create(
-      NvCodecV4L2CapturerConfig config);
+      const NvCodecV4L2CapturerConfig& config);
   NvCodecV4L2Capturer(const NvCodecV4L2CapturerConfig& config);
-
-  bool UseNativeBuffer() override;
 
  private:
   static rtc::scoped_refptr<V4L2VideoCapturer> Create(
       webrtc::VideoCaptureModule::DeviceInfo* device_info,
-      NvCodecV4L2CapturerConfig config,
+      const NvCodecV4L2CapturerConfig& config,
       size_t capture_device_index);
 
   void OnCaptured(uint8_t* data, uint32_t bytesused) override;
@@ -31,4 +30,6 @@ class NvCodecV4L2Capturer : public V4L2VideoCapturer {
   std::shared_ptr<NvCodecDecoderCuda> decoder_;
 };
 
-#endif  // NVCODEC_V4L2_CAPTURER_H_
+}  // namespace sora
+
+#endif
