+++ src/sora-cpp-sdk/src/hwenc_nvcodec/nvcodec_h264_encoder_cuda.cpp
@@ -1,4 +1,4 @@
-#include "fix_cuda_noinline_macro_error.h"
+#include "sora/fix_cuda_noinline_macro_error.h"
 
 #include "nvcodec_h264_encoder_cuda.h"
 
@@ -8,8 +8,10 @@
 #include <NvDecoder/NvDecoder.h>
 #include <NvEncoder/NvEncoderCuda.h>
 
-#include "cuda/cuda_context_cuda.h"
-#include "dyn/cuda.h"
+#include "../cuda_context_cuda.h"
+#include "sora/dyn/cuda.h"
+
+namespace sora {
 
 class NvCodecH264EncoderCudaImpl {
  public:
@@ -17,7 +19,7 @@ class NvCodecH264EncoderCudaImpl {
   ~NvCodecH264EncoderCudaImpl();
 
   void Copy(NvEncoder* nv_encoder, const void* ptr, int width, int height);
-  NvEncoder* CreateNvEncoder(int width, int height, bool use_native);
+  NvEncoder* CreateNvEncoder(int width, int height, bool is_nv12);
 
  private:
   std::shared_ptr<CudaContext> cuda_context_;
@@ -37,8 +39,8 @@ void NvCodecH264EncoderCuda::Copy(NvEncoder* nv_encoder,
 }
 NvEncoder* NvCodecH264EncoderCuda::CreateNvEncoder(int width,
                                                    int height,
-                                                   bool use_native) {
-  return impl_->CreateNvEncoder(width, height, use_native);
+                                                   bool is_nv12) {
+  return impl_->CreateNvEncoder(width, height, is_nv12);
 }
 
 void ShowEncoderCapability() {
@@ -61,68 +63,79 @@ void ShowEncoderCapability() {
     std::cout << "GPU " << iGpu << " - " << szDeviceName << std::endl
               << std::endl;
     std::cout
-        << "\tH264:\t\t" << "  "
+        << "\tH264:\t\t"
+        << "  "
         << (enc.GetCapabilityValue(NV_ENC_CODEC_H264_GUID,
                                    NV_ENC_CAPS_SUPPORTED_RATECONTROL_MODES)
                 ? "yes"
                 : "no")
         << std::endl
-        << "\tH264_444:\t" << "  "
+        << "\tH264_444:\t"
+        << "  "
         << (enc.GetCapabilityValue(NV_ENC_CODEC_H264_GUID,
                                    NV_ENC_CAPS_SUPPORT_YUV444_ENCODE)
                 ? "yes"
                 : "no")
         << std::endl
-        << "\tH264_ME:\t" << "  "
+        << "\tH264_ME:\t"
+        << "  "
         << (enc.GetCapabilityValue(NV_ENC_CODEC_H264_GUID,
                                    NV_ENC_CAPS_SUPPORT_MEONLY_MODE)
                 ? "yes"
                 : "no")
         << std::endl
-        << "\tH264_WxH:\t" << "  "
+        << "\tH264_WxH:\t"
+        << "  "
         << (enc.GetCapabilityValue(NV_ENC_CODEC_H264_GUID,
                                    NV_ENC_CAPS_WIDTH_MAX))
         << "*"
         << (enc.GetCapabilityValue(NV_ENC_CODEC_H264_GUID,
                                    NV_ENC_CAPS_HEIGHT_MAX))
         << std::endl
-        << "\tHEVC:\t\t" << "  "
+        << "\tHEVC:\t\t"
+        << "  "
         << (enc.GetCapabilityValue(NV_ENC_CODEC_HEVC_GUID,
                                    NV_ENC_CAPS_SUPPORTED_RATECONTROL_MODES)
                 ? "yes"
                 : "no")
         << std::endl
-        << "\tHEVC_Main10:\t" << "  "
+        << "\tHEVC_Main10:\t"
+        << "  "
         << (enc.GetCapabilityValue(NV_ENC_CODEC_HEVC_GUID,
                                    NV_ENC_CAPS_SUPPORT_10BIT_ENCODE)
                 ? "yes"
                 : "no")
         << std::endl
-        << "\tHEVC_Lossless:\t" << "  "
+        << "\tHEVC_Lossless:\t"
+        << "  "
         << (enc.GetCapabilityValue(NV_ENC_CODEC_HEVC_GUID,
                                    NV_ENC_CAPS_SUPPORT_LOSSLESS_ENCODE)
                 ? "yes"
                 : "no")
         << std::endl
-        << "\tHEVC_SAO:\t" << "  "
+        << "\tHEVC_SAO:\t"
+        << "  "
         << (enc.GetCapabilityValue(NV_ENC_CODEC_HEVC_GUID,
                                    NV_ENC_CAPS_SUPPORT_SAO)
                 ? "yes"
                 : "no")
         << std::endl
-        << "\tHEVC_444:\t" << "  "
+        << "\tHEVC_444:\t"
+        << "  "
         << (enc.GetCapabilityValue(NV_ENC_CODEC_HEVC_GUID,
                                    NV_ENC_CAPS_SUPPORT_YUV444_ENCODE)
                 ? "yes"
                 : "no")
         << std::endl
-        << "\tHEVC_ME:\t" << "  "
+        << "\tHEVC_ME:\t"
+        << "  "
         << (enc.GetCapabilityValue(NV_ENC_CODEC_HEVC_GUID,
                                    NV_ENC_CAPS_SUPPORT_MEONLY_MODE)
                 ? "yes"
                 : "no")
         << std::endl
-        << "\tHEVC_WxH:\t" << "  "
+        << "\tHEVC_WxH:\t"
+        << "  "
         << (enc.GetCapabilityValue(NV_ENC_CODEC_HEVC_GUID,
                                    NV_ENC_CAPS_WIDTH_MAX))
         << "*"
@@ -157,9 +170,11 @@ void NvCodecH264EncoderCudaImpl::Copy(NvEncoder* nv_encoder,
 }
 NvEncoder* NvCodecH264EncoderCudaImpl::CreateNvEncoder(int width,
                                                        int height,
-                                                       bool use_native) {
+                                                       bool is_nv12) {
   NV_ENC_BUFFER_FORMAT nvenc_format =
-      use_native ? NV_ENC_BUFFER_FORMAT_NV12 : NV_ENC_BUFFER_FORMAT_IYUV;
+      is_nv12 ? NV_ENC_BUFFER_FORMAT_NV12 : NV_ENC_BUFFER_FORMAT_IYUV;
   CUcontext context = GetCudaContext(cuda_context_);
   return new NvEncoderCuda(context, width, height, nvenc_format);
 }
+
+}  // namespace sora
