+++ src/sora-cpp-sdk/src/hwenc_nvcodec/nvcodec_video_decoder.cpp
@@ -1,6 +1,4 @@
-#include "fix_cuda_noinline_macro_error.h"
-
-#include "nvcodec_video_decoder.h"
+#include "sora/hwenc_nvcodec/nvcodec_video_decoder.h"
 
 // WebRTC
 #include <modules/video_coding/include/video_error_codes.h>
@@ -9,8 +7,13 @@
 #include <rtc_base/time_utils.h>
 #include <third_party/libyuv/include/libyuv/convert.h>
 
-#include "dyn/cuda.h"
-#include "dyn/nvcuvid.h"
+// NvCodec
+#include <NvDecoder/NvDecoder.h>
+
+#include "sora/dyn/cuda.h"
+#include "sora/dyn/nvcuvid.h"
+
+namespace sora {
 
 NvCodecVideoDecoder::NvCodecVideoDecoder(std::shared_ptr<CudaContext> ctx,
                                          CudaVideoCodec codec)
@@ -23,7 +26,12 @@ NvCodecVideoDecoder::~NvCodecVideoDecoder() {
   Release();
 }
 
-bool NvCodecVideoDecoder::IsSupported(CudaVideoCodec codec) {
+bool NvCodecVideoDecoder::IsSupported(std::shared_ptr<CudaContext> context,
+                                      CudaVideoCodec codec) {
+  if (context == nullptr) {
+    return false;
+  }
+
   // CUDA 周りのライブラリがロードできるか確認する
   if (!dyn::DynModule::Instance().IsLoadable(dyn::CUDA_SO)) {
     return false;
@@ -31,9 +39,17 @@ bool NvCodecVideoDecoder::IsSupported(CudaVideoCodec codec) {
   if (!dyn::DynModule::Instance().IsLoadable(dyn::NVCUVID_SO)) {
     return false;
   }
+  // 関数が存在するかチェックする
+  if (dyn::DynModule::Instance().GetFunc(dyn::CUDA_SO, "cuDeviceGetName") ==
+      nullptr) {
+    return false;
+  }
+  if (dyn::DynModule::Instance().GetFunc(dyn::NVCUVID_SO,
+                                         "cuvidMapVideoFrame") == nullptr) {
+    return false;
+  }
 
   try {
-    auto context = CudaContext::Create();
     auto p = new NvCodecDecoderCuda(context, codec);
     delete p;
     return true;
@@ -60,8 +76,13 @@ int32_t NvCodecVideoDecoder::Decode(const webrtc::EncodedImage& input_image,
   }
 
   uint8_t** frames = nullptr;
-  int frame_count =
-      decoder_->Decode(input_image.data(), (int)input_image.size());
+  int frame_count;
+  try {
+    frame_count = decoder_->Decode(input_image.data(), (int)input_image.size());
+  } catch (NVDECException& e) {
+    RTC_LOG(LS_ERROR) << e.getErrorString();
+    return WEBRTC_VIDEO_CODEC_ERROR;
+  }
   if (frame_count == 0) {
     return WEBRTC_VIDEO_CODEC_OK;
   }
@@ -75,7 +96,7 @@ int32_t NvCodecVideoDecoder::Decode(const webrtc::EncodedImage& input_image,
   uint32_t pts = input_image.RtpTimestamp();
 
   for (int i = 0; i < frame_count; i++) {
-    const auto* frame = decoder_->GetFrame();
+    auto* frame = decoder_->GetLockedFrame();
     // NV12 から I420 に変換
     rtc::scoped_refptr<webrtc::I420Buffer> i420_buffer =
         buffer_pool_.CreateI420Buffer(decoder_->GetWidth(),
@@ -95,9 +116,10 @@ int32_t NvCodecVideoDecoder::Decode(const webrtc::EncodedImage& input_image,
     decode_complete_callback_->Decoded(decoded_image, absl::nullopt,
                                        absl::nullopt);
 
-    // 次のフレームで縦横サイズが変わったときに追従するためのマジックコード
-    decoder_->setReconfigParams();
+    decoder_->UnlockFrame(frame);
   }
+  // 次のフレームで縦横サイズが変わったときに追従するためのマジックコード
+  decoder_->setReconfigParams();
 
   return WEBRTC_VIDEO_CODEC_OK;
 }
@@ -127,3 +149,5 @@ bool NvCodecVideoDecoder::InitNvCodec() {
 void NvCodecVideoDecoder::ReleaseNvCodec() {
   decoder_.reset();
 }
+
+}  // namespace sora
