+++ src/sora-cpp-sdk/src/hwenc_vpl/vaapi_utils.h
