+++ src/sora-cpp-sdk/src/hwenc_vpl/vpl_utils.h
