+++ src/sora-cpp-sdk/src/hwenc_vpl/vpl_video_encoder.cpp
